requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Integration
anthropic==0.7.7

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
xlrd==2.0.1

# Report Generation
reportlab==4.0.7
weasyprint==60.1
python-docx==1.1.0
Pillow==10.1.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Frontend
streamlit==1.28.2

# Database (optional for future use)
sqlalchemy==2.0.23

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
